Change all IDs to float64.
If you pass large-ish integers (123456789) through Go unmarshal/marshal
with a generic map[string]interface{}, it will be re-encoded as
1.23456789e+9. Re-decoding that into an uint promptly returns an
error...

+++ app.go
@@ -3,26 +3,26 @@ package podio
 import "fmt"
 
 type App struct {
-	Id       uint   `json:"app_id"`
-	Name     string `json:"name"`
-	ItemName string `json:"item_name"`
-	SpaceId  uint   `json:"space_id"`
+	Id       float64 `json:"app_id"`
+	Name     string  `json:"name"`
+	ItemName string  `json:"item_name"`
+	SpaceId  float64 `json:"space_id"`
 }
 
-func (client *Client) GetApps(space_id uint) (apps []App, err error) {
-	path := fmt.Sprintf("/app/space/%d?view=micro", space_id)
+func (client *Client) GetApps(space_id float64) (apps []App, err error) {
+	path := fmt.Sprintf("/app/space/%.0f?view=micro", space_id)
 	err = client.Request("GET", path, nil, nil, &apps)
 	return
 }
 
-func (client *Client) GetApp(id uint) (app *App, err error) {
-	path := fmt.Sprintf("/app/%d?view=micro", id)
+func (client *Client) GetApp(id float64) (app *App, err error) {
+	path := fmt.Sprintf("/app/%.0f?view=micro", id)
 	err = client.Request("GET", path, nil, nil, &app)
 	return
 }
 
-func (client *Client) GetAppBySpaceIdAndSlug(space_id uint, slug string) (app *App, err error) {
-	path := fmt.Sprintf("/app/space/%d/%s", space_id, slug)
+func (client *Client) GetAppBySpaceIdAndSlug(space_id float64, slug string) (app *App, err error) {
+	path := fmt.Sprintf("/app/space/%.0f/%s", space_id, slug)
 	err = client.Request("GET", path, nil, nil, &app)
 	return
 }
