+++ item.go
@@ -6,9 +6,9 @@ import (
 )
 
 type Item struct {
-	Id                 uint     `json:"item_id"`
+	Id                 float64  `json:"item_id"`
 	App                *App     `json:"app"`
-	AppItemId          uint     `json:"app_item_id"`
+	AppItemId          float64  `json:"app_item_id"`
 	FormattedAppItemId string   `json:"app_item_id_formatted"`
 	Link               string   `json:"link"`
 	Title              string   `json:"title"`
@@ -17,7 +17,7 @@ type Item struct {
 }
 
 type Field struct {
-	FieldId    uint     `json:"field_id"`
+	FieldId    float64  `json:"field_id"`
 	ExternalId string   `json:"external_id"`
 	Type       string   `json:"type"`
 	Label      string   `json:"label"`
@@ -26,7 +26,7 @@ type Field struct {
 
 func (f *Field) UnmarshalJSON(data []byte) error {
 	var fake struct {
-		FieldId    uint     `json:"field_id"`
+		FieldId    float64  `json:"field_id"`
 		ExternalId string   `json:"external_id"`
 		Type       string   `json:"type"`
 		Label      string   `json:"label"`
@@ -72,37 +72,37 @@ type Value struct {
 }
 
 type ItemList struct {
-	Filtered uint    `json:"filtered"`
-	Total    uint    `json:"total"`
+	Filtered float64 `json:"filtered"`
+	Total    float64 `json:"total"`
 	Items    []*Item `json:"items"`
 }
 
-func (client *Client) GetItems(app_id uint) (items *ItemList, err error) {
-	path := fmt.Sprintf("/item/app/%d/filter?fields=items.fields(files)", app_id)
+func (client *Client) GetItems(app_id float64) (items *ItemList, err error) {
+	path := fmt.Sprintf("/item/app/%.0f/filter?fields=items.fields(files)", app_id)
 	err = client.Request("POST", path, nil, nil, &items)
 	return
 }
 
-func (client *Client) GetItemByAppItemId(app_id uint, formatted_app_item_id string) (item *Item, err error) {
-	path := fmt.Sprintf("/app/%d/item/%s", app_id, formatted_app_item_id)
+func (client *Client) GetItemByAppItemId(app_id float64, formatted_app_item_id string) (item *Item, err error) {
+	path := fmt.Sprintf("/app/%.0f/item/%s", app_id, formatted_app_item_id)
 	err = client.Request("GET", path, nil, nil, &item)
 	return
 }
 
-func (client *Client) GetItemByExternalID(app_id uint, external_id string) (item *Item, err error) {
-	path := fmt.Sprintf("/item/app/%d/external_id/%s", app_id, external_id)
+func (client *Client) GetItemByExternalID(app_id float64, external_id string) (item *Item, err error) {
+	path := fmt.Sprintf("/item/app/%.0f/external_id/%s", app_id, external_id)
 	err = client.Request("GET", path, nil, nil, &item)
 	return
 }
 
-func (client *Client) GetItem(item_id uint) (item *Item, err error) {
-	path := fmt.Sprintf("/item/%d?fields=files", item_id)
+func (client *Client) GetItem(item_id float64) (item *Item, err error) {
+	path := fmt.Sprintf("/item/%.0f?fields=files", item_id)
 	err = client.Request("GET", path, nil, nil, &item)
 	return
 }
 
-func (client *Client) CreateItem(app_id uint, external_id string, fieldValues map[string]interface{}) (uint, error) {
-	path := fmt.Sprintf("/item/app/%d", app_id)
+func (client *Client) CreateItem(app_id float64, external_id string, fieldValues map[string]interface{}) (float64, error) {
+	path := fmt.Sprintf("/item/app/%.0f", app_id)
 	params := map[string]interface{}{
 		"fields": fieldValues,
 	}
@@ -112,15 +112,15 @@ func (client *Client) CreateItem(app_id uint, external_id string, fieldValues ma
 	}
 
 	rsp := &struct {
-		ItemId uint `json:"item_id"`
+		ItemId float64 `json:"item_id"`
 	}{}
 	err := client.RequestWithParams("POST", path, nil, params, rsp)
 
 	return rsp.ItemId, err
 }
 
-func (client *Client) UpdateItem(itemId uint, fieldValues map[string]interface{}) error {
-	path := fmt.Sprintf("/item/%d", itemId)
+func (client *Client) UpdateItem(itemId float64, fieldValues map[string]interface{}) error {
+	path := fmt.Sprintf("/item/%.0f", itemId)
 	params := map[string]interface{}{
 		"fields": fieldValues,
 	}
