+++ space.go
@@ -3,25 +3,25 @@ package podio
 import "fmt"
 
 type Space struct {
-	Id   uint   `json:"space_id"`
-	Slug string `json:"url_label"`
-	Name string `json:"name"`
+	Id   float64 `json:"space_id"`
+	Slug string  `json:"url_label"`
+	Name string  `json:"name"`
 }
 
-func (client *Client) GetSpaces(org_id uint) (spaces []Space, err error) {
-	path := fmt.Sprintf("/org/%d/space", org_id)
+func (client *Client) GetSpaces(org_id float64) (spaces []Space, err error) {
+	path := fmt.Sprintf("/org/%.0f/space", org_id)
 	err = client.Request("GET", path, nil, nil, &spaces)
 	return
 }
 
-func (client *Client) GetSpace(id uint) (space *Space, err error) {
-	path := fmt.Sprintf("/space/%d", id)
+func (client *Client) GetSpace(id float64) (space *Space, err error) {
+	path := fmt.Sprintf("/space/%.0f", id)
 	err = client.Request("GET", path, nil, nil, &space)
 	return
 }
 
-func (client *Client) GetSpaceByOrgIdAndSlug(org_id uint, slug string) (space *Space, err error) {
-	path := fmt.Sprintf("/space/org/%d/%s", org_id, slug)
+func (client *Client) GetSpaceByOrgIdAndSlug(org_id float64, slug string) (space *Space, err error) {
+	path := fmt.Sprintf("/space/org/%.0f/%s", org_id, slug)
 	err = client.Request("GET", path, nil, nil, &space)
 	return
 }
